Rename generate geodatabase (#1008)
