+++ java/generate-geodatabase-replica-from-feature-service/README.metadata.json
@@ -1,0 +1,40 @@
+{
+    "category": "Edit and Manage Data",
+    "description": "Generate a local geodatabase replica from an online feature service.",
+    "formal_name": "GenerateGeodatabaseReplicaFromFeatureService",
+    "ignore": false,
+    "images": [
+        "generate-geodatabase-replica-from-feature-service.png"
+    ],
+    "keywords": [
+        "disconnected",
+        "local geodatabase",
+        "offline",
+        "replica",
+        "sync",
+        "GenerateGeodatabaseJob",
+        "GenerateGeodatabaseParameters",
+        "Geodatabase",
+        "GeodatabaseSyncTask"
+    ],
+    "language": "java",
+    "provision_from": [
+        "https://arcgisruntime.maps.arcgis.com/home/item.html?id=e4a398afe9a945f3b0f4dca1e4faccb5"
+    ],
+    "provision_to": [
+        "/SanFrancisco.tpkx"
+    ],
+    "redirect_from": [
+        "/android/latest/sample-code/generate-geodatabase.htm"
+    ],
+    "relevant_apis": [
+        "GenerateGeodatabaseJob",
+        "GenerateGeodatabaseParameters",
+        "Geodatabase",
+        "GeodatabaseSyncTask"
+    ],
+    "snippets": [
+        "src/main/java/com/esri/arcgisruntime/sample/generategeodatabasereplicafromfeatureservice/MainActivity.java"
+    ],
+    "title": "Generate geodatabase replica from feature service"
+}
