+++ java/generate-geodatabase-replica-from-feature-service/src/main/res/drawable/preview_background.xml
