+++ java/generate-geodatabase-replica-from-feature-service/src/main/res/layout/activity_main.xml
