+++ java/generate-geodatabase-replica-from-feature-service/src/main/res/mipmap-hdpi/ic_launcher.png
