+++ java/generate-geodatabase-replica-from-feature-service/src/main/res/mipmap-xhdpi/ic_launcher.png
