+++ java/settings.gradle
@@ -84,7 +84,7 @@ include ':add-enc-exchange-set',
         ':find-route',
         ':find-service-area-interactive',
         ':format-coordinates',
-        ':generate-geodatabase',
+        ':generate-geodatabase-replica-from-feature-service',
         ':generate-offline-map',
         ':generate-offline-map-overrides',
         ':generate-offline-map-with-local-basemap',
