+++ kotlin/generate-geodatabase-replica-from-feature-service/proguard-rules.pro
