+++ kotlin/generate-geodatabase-replica-from-feature-service/src/main/res/drawable/ic_launcher_background.xml
