+++ kotlin/generate-geodatabase-replica-from-feature-service/src/main/res/mipmap-anydpi-v26/ic_launcher.xml
