+++ kotlin/generate-geodatabase-replica-from-feature-service/src/main/res/mipmap-mdpi/ic_launcher.png
