+++ kotlin/generate-geodatabase-replica-from-feature-service/src/main/res/mipmap-xxhdpi/ic_launcher.png
