+++ kotlin/generate-geodatabase-replica-from-feature-service/src/main/res/mipmap-xxxhdpi/ic_launcher.png
