+++ kotlin/generate-geodatabase-replica-from-feature-service/src/main/res/values/colors.xml
