--- kotlin/generate-geodatabase/README.metadata.json
@@ -1,35 +1,0 @@
-{
-    "category": "Edit and Manage Data",
-    "description": "Generate a local geodatabase from an online feature service.",
-    "formal_name": "GenerateGeodatabase",
-    "ignore": false,
-    "images": [
-        "generate-geodatabase.png"
-    ],
-    "keywords": [
-        "disconnected",
-        "local geodatabase",
-        "offline",
-        "sync"
-    ],
-    "language": "kotlin",
-    "provision_from": [
-        "https://arcgisruntime.maps.arcgis.com/home/item.html?id=e4a398afe9a945f3b0f4dca1e4faccb5"
-    ],
-    "provision_to": [
-        "/SanFrancisco.tpkx"
-    ],
-    "redirect_from": [
-        "/android/latest/sample-code/generate-geodatabase.htm"
-    ],
-    "relevant_apis": [
-        "GenerateGeodatabaseJob",
-        "GenerateGeodatabaseParameters",
-        "Geodatabase",
-        "GeodatabaseSyncTask"
-    ],
-    "snippets": [
-        "src/main/java/com/esri/arcgisruntime/sample/generategeodatabase/MainActivity.kt"
-    ],
-    "title": "Generate geodatabase"
-}
