--- kotlin/generate-geodatabase/build.gradle
@@ -1,30 +1,0 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-
-android {
-    compileSdkVersion rootProject.ext.compileSdkVersion
-
-    defaultConfig {
-        applicationId "com.esri.arcgisruntime.sample.generategeodatabase"
-        minSdkVersion rootProject.ext.minSdkVersion
-        targetSdkVersion rootProject.ext.targetSdkVersion
-        versionCode rootProject.ext.versionCode
-        versionName rootProject.ext.versionName
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-
-    buildFeatures {
-        viewBinding true
-    }
-}
-
-dependencies {
-    // arcgis-android & appcompat lib dependencies from rootProject build.gradle
-    implementation "androidx.constraintlayout:constraintlayout:$constraintLayoutVersion"
-}
