+++ kotlin/settings.gradle
@@ -69,7 +69,7 @@ include ':add-features-feature-service',
         ':filter-by-definition-expression-or-display-filter',
         ':find-address',
         ':find-route',
-        ':generate-geodatabase',
+        ':generate-geodatabase-replica-from-feature-service',
         ':generate-offline-map',
         ':geodesic-operations',
         ':group-layers',
